requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
django-countries==7.6.1
django-dramatiq==0.11.6
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
dramatiq==1.17.0
dramatiq-crontab==1.0.7
factory-boy==3.3.0
Faker==26.2.0
idna==3.7
phonenumbers==8.13.38
pika==1.3.2
pika-stubs==0.1.3
prometheus_client==0.20.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
tzlocal==5.2
urllib3==2.2.2
vine==5.1.0
